CONVERT modified to take a from and a to type argument.
This makes convert far more flexible and lets me indicate what type the
input should be considered as. Some ideas I have planned for this is an
easy way to convert a number to a list or an array of that number in
various bases like binary and hexidecimal.

The idea for the CONVERT generic method came to me from FSET, but I have
extended my version to be more specific about what type the input is meant
to be considered as. No source is taken from FSET, the only thing FSET
gave me here is the generic function name... something I have been
struggling with for at least a month now. Its too bad coerce is already
taken. ;)

You can find FSET's homepage at <http://common-lisp.net/project/fset/>.