This actually "assembles" in a very very limited sense!
This is based on some stuff in AoA, reference:

http://www.arl.wustl.edu/~lockwood/class/cs306/books/artofasm/Chapter_3/CH03-3.html#HEADING3-102

However! note that for some reason or another ndisasm is having a fit when
given mov ax, bx. Supposed to be 1100 0001. This is a part of Shift Grp,
not mov.

Next step is taking what was tested and perfected here (binary-slot-value
especially!) and using this information to start creating real opcodes
based on the stuff sdkmvx pushed in and the intel manuals.

This code if nothing else is useful to see a pratical example binary-data.