Define a class for push.
This is a crude hack until BINARY-DATA's new bug is fixed. Once the bug is
fixed I think we will be set to go as far as how the assembler is laid out
and so forth.

Note that in this commit we shadow CL:PUSH in nass.arch.x86iods to stick
with the assembler name.